backend/requirements.txt
Flask==2.0.3
python-dotenv==0.21.0
numpy==1.24.2
cellpose==2.2.2
Werkzeug==2.0.3
Pillow==9.5.0